breaking out decoderImpl code
