+++ Sources/AutomergeSwiftAdditions/Codable/Decoding/AutomergeDecoder.swift
@@ -33,70 +33,3 @@ public struct AutomergeDecoder {
         return try decoder.decode(T.self)
     }
 }
-
-@usableFromInline struct AutomergeDecoderImpl {
-    @usableFromInline let doc: Document
-    @usableFromInline let codingPath: [CodingKey]
-    @usableFromInline let userInfo: [CodingUserInfoKey: Any]
-
-    @usableFromInline let automergeValue: AutomergeValue
-
-    @inlinable init(
-        doc: Document,
-        userInfo: [CodingUserInfoKey: Any],
-        from automergeValue: AutomergeValue,
-        codingPath: [CodingKey]
-    ) {
-        self.doc = doc
-        self.userInfo = userInfo
-        self.codingPath = codingPath
-        self.automergeValue = automergeValue
-    }
-
-    @inlinable public func decode<T: Decodable>(_: T.Type) throws -> T {
-        try T(from: self)
-    }
-}
-
-extension AutomergeDecoderImpl: Decoder {
-    @usableFromInline func container<Key>(keyedBy _: Key.Type) throws ->
-        KeyedDecodingContainer<Key> where Key: CodingKey
-    {
-        guard case let .object(dictionary) = self.automergeValue else {
-            throw DecodingError.typeMismatch([String: AutomergeValue].self, DecodingError.Context(
-                codingPath: self.codingPath,
-                debugDescription: "Expected to decode \([String: AutomergeValue].self) but found \(self.automergeValue.debugDataTypeDescription) instead."
-            ))
-        }
-        // dictionary: [String: AutomergeValue]
-        let container = AutomergeKeyedDecodingContainer<Key>(
-            impl: self,
-            codingPath: codingPath,
-            dictionary: dictionary
-        )
-        return KeyedDecodingContainer(container)
-    }
-
-    @usableFromInline func unkeyedContainer() throws -> UnkeyedDecodingContainer {
-        guard case let .array(array) = self.automergeValue else {
-            throw DecodingError.typeMismatch([AutomergeValue].self, DecodingError.Context(
-                codingPath: self.codingPath,
-                debugDescription: "Expected to decode \([AutomergeValue].self) but found \(self.automergeValue.debugDataTypeDescription) instead."
-            ))
-        }
-
-        return AutomergeUnkeyedDecodingContainer(
-            impl: self,
-            codingPath: self.codingPath,
-            array: array
-        )
-    }
-
-    @usableFromInline func singleValueContainer() throws -> SingleValueDecodingContainer {
-        AutomergeSingleValueDecodingContainer(
-            impl: self,
-            codingPath: self.codingPath,
-            automergeValue: self.automergeValue
-        )
-    }
-}
